+++ src/main/java/fr/inrae/act/bagap/chloe/window/analysis/map/TinyMapLandscapeMetricAnalysis.java
@@ -29,7 +29,7 @@ public class TinyMapLandscapeMetricAnalysis extends MapLandscapeMetricAnalysis {
 		coverage().dispose();
 		
 		// gestion des sorties
-		kernel().setOutDatas(new float[nbValues()]);
+		kernel().setOutDatas(new double[nbValues()]);
 		kernel().outDatas()[0] = 1; // filtre ok
 		
 		// initialisation du comptage
