+++ src/main/java/fr/inrae/act/bagap/chloe/window/analysis/selected/TinySelectedLandscapeMetricAnalysis.java
@@ -33,9 +33,9 @@ public class TinySelectedLandscapeMetricAnalysis extends SelectedLandscapeMetric
 		manageInDatas(roi);
 		
 		// gestion des sorties
-		Map<Pixel, float[]> outDatas = new HashMap<Pixel, float[]>();
+		Map<Pixel, double[]> outDatas = new HashMap<Pixel, double[]>();
 		for(Pixel p : pixels()){
-			outDatas.put(p, new float[nbValues()]);
+			outDatas.put(p, new double[nbValues()]);
 		}
 		kernel().setOutDatas(outDatas);
